- Replaced all usages of Assert across tests/docs with Shouldly assertions, including collection/predicate patterns and failure scenarios to align with Shouldly’s readable syntax and     richer messages (docs.shouldly.org (https://docs.shouldly.org/?utm_source=openai)).   - Fixed predicate-based checks that previously relied on Assert.All/Contains by adopting Shouldly helpers or LINQ combos that match the framework’s recommended style     (shouldly.readthedocs.io (https://shouldly.readthedocs.io/en/latest/?utm_source=openai)).   - Ensured Shouldly-friendly comparisons for numeric/nullable/task-returning expressions (e.g., awaited WaitToReadAsync, explicit byte casts) to compile cleanly under .NET 10.0     targets.   - Updated README/CONTRIBUTING snippets to demonstrate Shouldly patterns so future contributors

+++ tests/Hugo.UnitTests/DeterministicRecordTests.cs
@@ -29,6 +29,6 @@ public class DeterministicRecordTests
 
         payload[0] = 42;
 
-        record.Payload.Span[0].ShouldBe(1);
+        record.Payload.Span[0].ShouldBe((byte)1);
     }
 }
